Refact: device data structures moved to relevant module (and updated includes in modules depending on it).
