+++ src/adf_file.c
@@ -38,6 +38,7 @@
 #include"adf_bitm.h"
 #include"adf_cache.h"
 #include "adf_env.h"
+#include "adf_dev.h"
 
 
 // debugging
