+++ src/adf_str.h
@@ -35,41 +35,6 @@
 #include"adf_err.h"
 
 
-
-struct Partition {
-    int32_t startCyl;
-    int32_t lenCyl;
-    char* volName;
-    int volType;
-};
-
-/* ----- DEVICES ----- */
-
-#define DEVTYPE_FLOPDD 		1
-#define DEVTYPE_FLOPHD 		2
-#define DEVTYPE_HARDDISK 	3
-#define DEVTYPE_HARDFILE 	4
-
-struct Device {
-    int devType;               /* see below */
-    BOOL readOnly;
-    int32_t size;                 /* in bytes */
-
-    int nVol;                  /* partitions */
-    struct Volume** volList;  
-	
-    int32_t cylinders;            /* geometry */
-    int32_t heads;
-    int32_t sectors;
-
-    BOOL isNativeDev;
-    union {
-        void *nativeDev;
-        FILE *fd;
-    };
-};
-
-
 /* ----- FILE ----- */
 
 struct File {
